requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pillow
torch
numpy
jinja2
requests
httpx